requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
awsebcli==3.20.3
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.11.1
bidict==0.22.0
boto3==1.25.1
botocore==1.28.1
bs4==0.0.1
cement==2.8.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
decorator==4.4.2
distlib==0.3.4
Django==3.2.4
django-cors-headers==3.12.0
django-embed-video==1.4.1
django-filter==21.1
django-storages==1.13.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
filelock==3.4.2
future==0.16.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
imageio==2.16.1
imageio-ffmpeg==0.4.5
incremental==21.3.0
iniconfig==1.1.1
jmespath==0.10.0
moviepy==1.0.3
numpy==1.22.2
oauthlib==3.2.1
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.4.1
pluggy==0.13.1
proglog==0.1.9
psycopg2==2.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.1.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2021.1
pywin32==304
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
s3transfer==0.6.0
semantic-version==2.8.5
service-identity==21.1.0
six==1.14.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
termcolor==1.1.0
toml==0.10.2
tqdm==4.63.0
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2021.5
urllib3==1.26.12
validate-email==1.3
virtualenv==20.10.0
wcwidth==0.1.9
wget==3.2
whitenoise==5.2.0
zope.interface==5.4.0
